# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ContaNova ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /proveedores:
    get:
      operationId: listProveedores
      tags: [proveedores]
      summary: Listar proveedores
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de proveedores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proveedor"
    post:
      operationId: createProveedor
      tags: [proveedores]
      summary: Crear proveedor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProveedorInput"
      responses:
        "201":
          description: Proveedor creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proveedor"
  /proveedores/{id}:
    get:
      operationId: getProveedor
      tags: [proveedores]
      summary: Obtener proveedor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proveedor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proveedor"
    patch:
      operationId: updateProveedor
      tags: [proveedores]
      summary: Actualizar proveedor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProveedorUpdate"
      responses:
        "200":
          description: Proveedor actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proveedor"
    delete:
      operationId: deleteProveedor
      tags: [proveedores]
      summary: Eliminar proveedor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Eliminado

  # ─── PRODUCTOS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Proveedor:
      type: object
      required:
        - id
        - nombre
        - tipoDocumento
        - numeroDocumento
        - activo
        - createdAt
      properties:
        id:
          type: integer
        nombre:
          type: string
        tipoDocumento:
          type: string
        numeroDocumento:
          type: string
        email:
          type: ["string", "null"]
        telefono:
          type: ["string", "null"]
        direccion:
          type: ["string", "null"]
        ciudad:
          type: ["string", "null"]
        activo:
          type: boolean
        createdAt:
          type: string
    ProveedorInput:
      type: object
      required: [nombre, tipoDocumento, numeroDocumento]
      properties:
        nombre:
          type: string
          minLength: 1
        tipoDocumento:
          type: string
        numeroDocumento:
          type: string
        email:
          type: string
        telefono:
          type: string
        direccion:
          type: string
        ciudad:
          type: string
    ProveedorUpdate:
      type: object
      properties:
        nombre:
          type: string
        tipoDocumento:
          type: string
        numeroDocumento:
          type: string
        email:
          type: string
        telefono:
          type: string
        direccion:
          type: string
        ciudad:
          type: string
        activo:
          type: boolean

    # ─── PRODUCTOS ────────────────────────────────────────────────────────────
...